backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1
python-multipart==0.0.6
celery==5.3.1
redis==4.5.5